[#36] class 블록 동적 생성 (#109)
* ✨ feat: 커스텀 툴박스 돔 조작으로 생성

* ✨ feat: alert 창을 이용한 css 블록 동적 생성

* 🔨 refactor: 빈 값 입력시 거절 alert 추가

* 🙀 chore: 함수 위치 변경

* 🔨 refactor: 파일 분리

* 🙀 chore: fsd 구조에 맞게 캡슐화

* 🙀 chore: 시맨틱 태그 사용 + ifram에 title추가

* 🔨 refactor: 프리뷰 부분 컴포넌트화로 분리

* 🙀 chore: import 설정 변경

* 🐛 fix: ci/cd 에러 수정
