+++ apps/client/src/widgets/workspace/blockly/htmBlockContents.ts
@@ -1,0 +1,26 @@
+export const htmBlockContents = [
+  {
+    kind: 'block',
+    type: 'html',
+  },
+  {
+    kind: 'block',
+    type: 'head',
+  },
+  {
+    kind: 'block',
+    type: 'body',
+  },
+  {
+    kind: 'block',
+    type: 'p',
+  },
+  {
+    kind: 'block',
+    type: 'button',
+  },
+  {
+    kind: 'block',
+    type: 'text',
+  },
+];
