+++ apps/client/src/widgets/workspace/blockly/toolboxConfig2.ts
@@ -1,0 +1,4 @@
+export const toolboxConfig2 = {
+  kind: 'categoryToolbox',
+  contents: [],
+};
